Move Workers registration to the file with all their code (like Perforce plugin)
IPlasticSourceControlWorker::RegisterWorkers(&Provider)
