+++ Source/PlasticSourceControl/Private/PlasticSourceControlModule.h
@@ -7,7 +7,7 @@
 #include "PlasticSourceControlProvider.h"
 
 /**
- * PlasticSourceControl is a simple Plastic SCM Source Control Plugin for Unreal Engine
+ * PlasticSourceControl is the official Plastic SCM Source Control Plugin for Unreal Engine
  *
  * Written and contributed by Sebastien Rombauts (sebastien.rombauts@gmail.com) for Codice Software
  */
